==> analysis/00_prep_file.Rmd <==

```{r}
library(tidyverse)
library(here)
library(caret)
library(ggthemes)

bd <- read_csv(here("data/behavioral_data/trimmed_lt_d.csv"))

resnet_sd <- read_csv(here("data/model_sim_data/resnet50.csv"))
eig_sd <- read_csv(here("data/model_sim_data/eig.csv"))
nonoise_sd <- read_csv(here("data/model_sim_data/nonoise.csv"))
nolearning_sd <- read_csv(here("data/model_sim_data/nolearning.csv"))
gen_sd <- read_csv(here("data/model_sim_data/eig_gen.csv"))

trial_info <- read_csv(here("data/model_sim_data/sim_info/trial_info_graded_dishab.csv"))
param_info <- read_csv(here("data/model_sim_data/sim_info/eig_selfpaced.csv"))

```


# behavioral dataset

## participants n 
```{r}
participant_n <- (bd %>% distinct(prolific_id) %>% count())$n
saveRDS(participant_n, here("cached_data/writing_cache/participant_n.Rds"))
```


```{r}
library(lme4)

bd <- bd %>% 
  mutate(is_first_trial = trial_number == 1) %>% 
  mutate(violation_type_with_background = case_when(
    trial_type == "background" ~ "background", 
    TRUE ~ violation_type
  )) %>% 
  mutate(
    animacy = case_when(
    grepl("inanimate", background_type) ~ "inanimate", 
    TRUE ~ "animate"), 
    
    number = case_when(
    grepl("single", background_type) ~ "single", 
    TRUE ~ "pair"
  ),
  
    pose = case_when(
    grepl("left", background_type) ~ "left", 
    TRUE ~ "right"
  )
    )
  




m <- lmerTest::lmer(
  log(total_rt) ~ trial_number * is_first_trial + 
    trial_number * is_first_trial * number + 
    trial_number * is_first_trial * pose + 
    trial_number * is_first_trial * animacy + 
    trial_number * is_first_trial * violation_type_with_background + log(block_number) +  (1|prolific_id), 
  data = bd
)



m %>% summary()
```

```{r}
bd
```

```{r}
trial_number = c(1, 2, 3, 4, 5, 6)
number = c("pair", "single")
animacy = c("animate", "inanimate")
pose = c("left", "righ")
violation_type = c("background")
block_number = seq(1:24)


background_trials = crossing(
  trial_number, 
  number, 
  animacy, 
  pose, 
  violation_type, 
  block_number
)

trial_number = c(2, 4, 6)
violation_type = c("animacy", "identity", "number", "pose")


deviant_trials = crossing(
  trial_number, 
  number, 
  animacy, 
  pose, 
  violation_type, 
  block_number
)


new_data <- bind_rows(background_trials, deviant_trials) %>% 
  rename(violation_type_with_background = violation_type) %>% 
  mutate(is_first_trial = trial_number == 1, 
         prolific_id = 1)


```


```{r}

new_data = bd %>% 
  distinct(trial_number, is_first_trial, violation_type_with_background, number, pose, animacy,violation_type_with_background, block_number) %>% 
  mutate(prolific_id = 1)
```


```{r}

new_data$pred = stats::predict(m, re.form = NA, newdata = new_data)
#new_data$pred <- predict(m, new_data = new_data)


```

```{r}
p1 <- new_data %>% 
    ggplot(aes(x = trial_number, y = pred, color = violation_type_with_background)) + scale_y_log10() +
   stat_summary(fun.data = "mean_cl_boot", position = position_dodge(width = .2)) + 
    stat_summary(fun.data = "mean_cl_boot", geom = "line", position = position_dodge(width = .2)) + 
  theme_few() + 
  theme(legend.position = "none")

p2 <- bd %>% 
  
  ggplot(aes(x = trial_number, y = log(total_rt), color = violation_type_with_background)) + scale_y_log10() +
   stat_summary(fun.data = "mean_cl_boot", position = position_dodge(width = .2)) + 
    stat_summary(fun.data = "mean_cl_boot", geom = "line", position = position_dodge(width = .2)) + 
  theme_few() 

library(patchwork)
p1 + p2
```


```{r}


beh_p <-  bd %>% 
  
  ggplot(aes(x = trial_number, y = log(total_rt), color = violation_type_with_background))  +
   stat_summary(fun.data = "mean_cl_boot", position = position_dodge(width = .2)) + 
  stat_summary(data = new_data, 
               fun.data = "mean_cl_boot", 
               aes(x = trial_number, y = pred, color = violation_type_with_background),
               geom = "line", position = position_dodge(width = .2)) + 
  theme_few() + 
  theme(legend.position = "left") 
beh_p

```


```{r}
source(here('analysis/helper/summarize_behavioral_data.R'))

raw_d <- read_csv(here("data/behavioral_data/trimmed_lt_d.csv"))
eig_sim <- read_csv(here("data/model_sim_data/eig_summarised.csv"))
eig_fit_res <- readRDS(here("cached_data/eig_aligned_fit_res.Rds"))
get_best_fit_res <- function(raw_d, fit_res, sim_res){
  best_fit_id = (fit_res %>% arrange(-mean_rsquared) %>% head(1))$param_id
  best_fit_res <- sim_res %>% filter(param_id == best_fit_id) 
  
  data = summarize_behavioral_data(raw_d) %>% ungroup() %>% left_join(best_fit_res %>% ungroup(), by = c("trial_number", "trial_type"))

  fitted_stats = colf_nlxb(mean_lt ~ mean_sample, data = data, lower = c(-Inf, 0.0000001))
  sample_slope = fitted_stats$coefficients["param_mean_sample"]
  sample_intercept = fitted_stats$coefficients["param_X.Intercept."]
    
    test_scaled <- data %>% 
      mutate(scaled_samples = mean_sample * sample_slope  + sample_intercept, 
             scaled_ub = ub_sample * sample_slope + sample_intercept, 
             scaled_lb = lb_sample * sample_slope + sample_intercept)
  
  return(test_scaled)
}


eig_best_fit <- get_best_fit_res(raw_d, eig_fit_res, eig_sim)

eig_search_plot <- eig_best_fit %>% 
  ggplot(aes(x = trial_number, y = log(scaled_samples), ymin = log(scaled_lb), ymax = log(scaled_ub), color  = trial_type)) + 
  #scale_y_log10() + 
  geom_pointrange(position = position_dodge(width = .2)) + 
  geom_line()+
  theme_few() + 
  theme(legend.position = "none") + 
  xlab("") + 
  ylab("Mean Samples") + 
  ggtitle("EIG - Best Fit") + 
  theme(plot.title = element_text(hjust = 0.5)) 
```

```{r}
beh_p + eig_search_plot 
```


```{r}
bd %>% 
  filter(trial_type != "background") %>% 
  group_by(violation_type) %>% 
  summarise(
    mean = mean(total_rt)
  )
```

## Visualizing 

```{r}
bd %>% 
  mutate(log_rt = log(total_rt)) %>% 
  pivot_longer(cols = c(log_rt, pred), 
               values_to = "val", 
               names_to = "val_type") %>% 
  mutate(trial_type = case_when(
    trial_type == "background" ~ "background", 
    block_type == "deviant_block" & trial_type == "deviant" ~ violation_type
  )) %>% 
  ggplot(aes(x = trial_number, y = val, color = trial_type)) + scale_y_log10() +
   stat_summary(fun.data = "mean_cl_boot", position = position_dodge(width = .2)) + 
    stat_summary(fun.data = "mean_cl_boot", geom = "line", position = position_dodge(width = .2)) + 
  theme_few() + 
  facet_grid(block_number~val_type)
  
```

## Running Model 

```{r}

```


## Full dataset summary (for parameter interpretation)

```{r}
bd_fit_summary <- bd %>% 
  mutate(trial_type = case_when(
    trial_type == "background" ~ "background", 
    block_type == "deviant_block" & trial_type == "deviant" ~ violation_type
  )) %>% 
   group_by(trial_number, trial_type) %>% 
  summarise(mean_lt = mean(total_rt), 
            mean_pred_lt = mean(pred_lt)) 

bd_fit_summary %>% write_csv(here("data/behavioral_data/bd_fit.csv"))

# %>%
#  ggplot(aes(x = trial_number, y = mean_lt, color = trial_type)) +
#  scale_y_log10() +
#  geom_point() +
#  geom_line(aes(group = trial_type)) +
#  theme_few()
```


# Simulation results 

## Visualizing / Sanity check 

```{r}
eig_summarized %>% 
  mutate(trial_type_overall = if_else(trial_type == "background", "background", "deviant")) %>% 
  ggplot(aes(x = trial_number, y = mean_sample, color = trial_type_overall)) + 
  stat_summary(fun.data = "mean_cl_boot", position = position_dodge(width = .2)) 


eig_summarized %>% 
   filter(trial_type != "background") %>% 
  group_by(param_id, trial_type) %>% 
  summarise(mean_sample  = mean(mean_sample)) %>% 
  ggplot(aes(x = trial_type, y = mean_sample, color = trial_type)) + 
  #geom_point(position = position_dodge(width  = .2), alpha = .3) + 
  stat_summary(fun.data = "mean_cl_boot", position = position_dodge(width = .2)) 
  
```

## summarize & merge sim results 

def pooled_sd(group):
    n = 20  # Size of each subset
    ssd = np.sum((n - 1) * group['n_samples_sd']**2)
    count = group['n_samples_sd'].count()
    return np.sqrt(np.sum((n - 1) * group['n_samples_sd']**2) / (n * count - count))
    



```{r}


# Function to calculate pooled standard deviation
pooled_sd <- function(sd, n) {
  sqrt(sum((n - 1) * sd^2) / sum(n - 1))
}

#function c

#ssd = np.sum((n - 1) * group['n_samples_sd']**2)
#count = group['n_samples_sd'].count()
#return np.sqrt(ssd / (n * count - count))

eig_summarized <- trial_info %>% 
  select(-test_type) %>% 
  distinct(trial_id, stim_id, fam_duration, violation_type) %>% 
  right_join(eig_sd, by = c("trial_id", "stim_id")) %>% 
  mutate(trial_number = index + 1) %>% 
  select(param_id, trial_number, fam_duration, violation_type, sample_n_mean, sample_n_std) %>% 
  mutate(trial_type = case_when(
    violation_type == "background" ~ "background", 
    violation_type == "identity" & (trial_number == fam_duration + 1) ~ "identity", 
    violation_type == "number" & (trial_number == fam_duration + 1) ~ "number", 
    violation_type == "pose" & (trial_number == fam_duration + 1) ~ "pose", 
    violation_type == "animacy" & (trial_number == fam_duration + 1) ~ "animacy", 
    TRUE ~ "background"
  )) %>% 
  select(param_id, trial_type, trial_number, sample_n_mean, sample_n_std) %>% 
  group_by(param_id, trial_type, trial_number) %>% 
  summarise(
    count = n(),
    #ssd = sum((20-1) * (sample_n_std ** 2)),
    #sd_sample = (ssd / 20 * count - count) ** .5, 
    sd_sample = pooled_sd(sample_n_std, 20),
    mean_sample = mean(sample_n_mean), 
    ub_sample = mean_sample + 1.96 * (sd_sample /((400 * count) ** .5)),
    lb_sample = mean_sample - 1.96 * (sd_sample /((400 * count) ** .5)),) %>% # missing SD here!
  ungroup() 
  
write_csv(eig_summarized, here("data/model_sim_data/eig_summarised.csv"))
  
eig_summarized %>%   
   ggplot(aes(x = trial_number, y = mean_sample, color = trial_type)) +
 scale_y_log10() +
 geom_point() +
 geom_line(aes(group = trial_type)) +
 theme_few()
  

 
```

## permuted trial info for eig 

```{r}


permuted_summarized_material <- trial_info %>% 
  select(-test_type) %>% 
  distinct(trial_id, stim_id, fam_duration, violation_type) %>% 
  right_join(eig_sd, by = c("trial_id", "stim_id")) %>% 
  mutate(trial_number = index + 1) %>% 
  select(param_id, trial_number, fam_duration, violation_type, sample_n_mean, sample_n_std) %>% 
  mutate(trial_type = case_when(
    violation_type == "background" ~ "background", 
    violation_type == "identity" & (trial_number == fam_duration + 1) ~ "identity", 
    violation_type == "number" & (trial_number == fam_duration + 1) ~ "number", 
    violation_type == "pose" & (trial_number == fam_duration + 1) ~ "pose", 
    violation_type == "animacy" & (trial_number == fam_duration + 1) ~ "animacy", 
    TRUE ~ "background"
  )) 



permuted_shuffle <- transform(permuted_summarized_material, trial_type = sample(trial_type))

permuted_trial_info <- permuted_shuffle %>% 
  select(param_id, trial_type, trial_number, sample_n_mean, sample_n_std) %>% 
  group_by(param_id, trial_type, trial_number) %>% 
  summarise(
    count = n(),
    #ssd = sum((20-1) * (sample_n_std ** 2)),
    #sd_sample = (ssd / 20 * count - count) ** .5, 
    sd_sample = pooled_sd(sample_n_std, 20),
    mean_sample = mean(sample_n_mean), 
    ub_sample = mean_sample + 1.96 * (sd_sample /((400 * count) ** .5)),
    lb_sample = mean_sample - 1.96 * (sd_sample /((400 * count) ** .5)),) %>% # missing SD here!
  ungroup() 

write_csv(permuted_trial_info, here("data/model_sim_data/permuted_eig_summarised.csv"))


permuted_trial_info %>%   
   ggplot(aes(x = trial_number, y = mean_sample, color = trial_type)) +
 scale_y_log10() +
stat_summary(fun.data = "mean_cl_boot") + 
 theme_few()
  

```



## no noise

```{r}
nonoise_summarized <- trial_info %>% 
  select(-test_type) %>% 
  distinct(trial_id, stim_id, fam_duration, violation_type) %>% 
  right_join(nonoise_sd, by = c("trial_id", "stim_id")) %>% 
  mutate(trial_number = index + 1) %>% 
  select(param_id, trial_number, fam_duration, violation_type, sample_n_mean, sample_n_std) %>% 
  mutate(trial_type = case_when(
    violation_type == "background" ~ "background", 
    violation_type == "identity" & (trial_number == fam_duration + 1) ~ "identity", 
    violation_type == "number" & (trial_number == fam_duration + 1) ~ "number", 
    violation_type == "pose" & (trial_number == fam_duration + 1) ~ "pose", 
    violation_type == "animacy" & (trial_number == fam_duration + 1) ~ "animacy", 
    TRUE ~ "background"
  )) %>% 
  select(param_id, trial_type, trial_number, sample_n_mean, sample_n_std) %>% 
  group_by(param_id, trial_type, trial_number) %>% 
  summarise(
    count = n(),
    #ssd = sum((20-1) * (sample_n_std ** 2)),
    #sd_sample = (ssd / 20 * count - count) ** .5, 
    sd_sample = pooled_sd(sample_n_std, 20),
    mean_sample = mean(sample_n_mean), 
    ub_sample = mean_sample + 1.96 * (sd_sample /((400 * count) ** .5)),
    lb_sample = mean_sample - 1.96 * (sd_sample /((400 * count) ** .5)),) %>% # missing SD here!
  ungroup() 
  
write_csv(nonoise_summarized, here("data/model_sim_data/nonoise_summarised.csv"))

nonoise_summarized %>%   
   ggplot(aes(x = trial_number, y = mean_sample, color = trial_type)) +
 scale_y_log10() +
 stat_summary(fun.data = "mean_cl_boot") 
 theme_few()
  
```


## no learning


```{r}
nolearning_summarised <- trial_info %>% 
  select(-test_type) %>% 
  distinct(trial_id, stim_id, fam_duration, violation_type) %>% 
  right_join(nolearning_sd, by = c("trial_id", "stim_id")) %>% 
  mutate(trial_number = index + 1) %>% 
  select(param_id, trial_number, fam_duration, violation_type, sample_n_mean, sample_n_std) %>% 
  mutate(trial_type = case_when(
    violation_type == "background" ~ "background", 
    violation_type == "identity" & (trial_number == fam_duration + 1) ~ "identity", 
    violation_type == "number" & (trial_number == fam_duration + 1) ~ "number", 
    violation_type == "pose" & (trial_number == fam_duration + 1) ~ "pose", 
    violation_type == "animacy" & (trial_number == fam_duration + 1) ~ "animacy", 
    TRUE ~ "background"
  )) %>% 
  select(param_id, trial_type, trial_number, sample_n_mean, sample_n_std) %>% 
  group_by(param_id, trial_type, trial_number) %>% 
  summarise(
    count = n(),
    #ssd = sum((20-1) * (sample_n_std ** 2)),
    #sd_sample = (ssd / 20 * count - count) ** .5, 
    sd_sample = pooled_sd(sample_n_std, 20),
    mean_sample = mean(sample_n_mean), 
    ub_sample = mean_sample + 1.96 * (sd_sample /((400 * count) ** .5)),
    lb_sample = mean_sample - 1.96 * (sd_sample /((400 * count) ** .5)),) %>% # missing SD here!
  ungroup() 
  
write_csv(nolearning_summarised, here("data/model_sim_data/nolearning_summarised.csv"))

nolearning_summarised %>%   
   ggplot(aes(x = trial_number, y = mean_sample, color = trial_type)) +
 scale_y_log10() +
 stat_summary(fun.data = "mean_cl_boot") 
```


```{r}
 trial_info %>% 
  select(-test_type) %>% 
  distinct(trial_id, stim_id, fam_duration, violation_type) %>% 
  right_join(eig_sd, by = c("trial_id", "stim_id")) %>% 
  mutate(trial_number = index + 1) %>% 
  select(param_id, trial_number, fam_duration, violation_type, sample_n_mean, sample_n_std) %>% 
  mutate(trial_type = case_when(
    violation_type == "background" ~ "background", 
    violation_type == "identity" & (trial_number == fam_duration + 1) ~ "identity", 
    violation_type == "number" & (trial_number == fam_duration + 1) ~ "number", 
    violation_type == "pose" & (trial_number == fam_duration + 1) ~ "pose", 
    violation_type == "animacy" & (trial_number == fam_duration + 1) ~ "animacy", 
    TRUE ~ "background"
  )) %>% 
  select(param_id, trial_type, trial_number, sample_n_mean, sample_n_std) %>% 
  group_by(param_id, trial_type, trial_number)
  
```


## generalization

```{r}

gen_summarised <- trial_info %>% 
  select(-test_type) %>% 
  distinct(trial_id, stim_id, fam_duration, violation_type) %>% 
  right_join(gen_sd, by = c("trial_id", "stim_id")) %>% 
  mutate(trial_number = index + 1) %>% 
  select(param_id, trial_number, fam_duration, violation_type, sample_n_mean, sample_n_std) %>% 
  mutate(trial_type = case_when(
    violation_type == "background" ~ "background", 
    violation_type == "identity" & (trial_number == fam_duration + 1) ~ "identity", 
    violation_type == "number" & (trial_number == fam_duration + 1) ~ "number", 
    violation_type == "pose" & (trial_number == fam_duration + 1) ~ "pose", 
    violation_type == "animacy" & (trial_number == fam_duration + 1) ~ "animacy", 
    TRUE ~ "background"
  )) %>% 
  select(param_id, trial_type, trial_number, sample_n_mean, sample_n_std) %>% 
  group_by(param_id, trial_type, trial_number) %>% 
  summarise(
    count = n(),
    #ssd = sum((20-1) * (sample_n_std ** 2)),
    #sd_sample = (ssd / 20 * count - count) ** .5, 
    sd_sample = pooled_sd(sample_n_std, 20),
    mean_sample = mean(sample_n_mean), 
    ub_sample = mean_sample + 1.96 * (sd_sample /((400 * count) ** .5)),
    lb_sample = mean_sample - 1.96 * (sd_sample /((400 * count) ** .5)),) %>% # missing SD here!
  ungroup() 
  
write_csv(gen_summarised, here("data/model_sim_data/eig_gen_summarised.csv"))

gen_summarised %>%   
   ggplot(aes(x = trial_number, y = mean_sample, color = trial_type)) +
 scale_y_log10() +
 stat_summary(fun.data = "mean_cl_boot") 
 theme_few()

```


## resnet 50

```{r}

resnet_summarised <- trial_info %>% 
  select(-test_type) %>% 
  distinct(trial_id, stim_id, fam_duration, violation_type) %>% 
  right_join(resnet_sd, by = c("trial_id", "stim_id")) %>% 
  mutate(trial_number = index + 1) %>% 
  select(param_id, trial_number, fam_duration, violation_type, sample_n_mean, sample_n_std) %>% 
  mutate(trial_type = case_when(
    violation_type == "background" ~ "background", 
    violation_type == "identity" & (trial_number == fam_duration + 1) ~ "identity", 
    violation_type == "number" & (trial_number == fam_duration + 1) ~ "number", 
    violation_type == "pose" & (trial_number == fam_duration + 1) ~ "pose", 
    violation_type == "animacy" & (trial_number == fam_duration + 1) ~ "animacy", 
    TRUE ~ "background"
  )) %>% 
  select(param_id, trial_type, trial_number, sample_n_mean, sample_n_std) %>% 
  group_by(param_id, trial_type, trial_number) %>% 
  summarise(
    count = n(),
    #ssd = sum((20-1) * (sample_n_std ** 2)),
    #sd_sample = (ssd / 20 * count - count) ** .5, 
    sd_sample = pooled_sd(sample_n_std, 20),
    mean_sample = mean(sample_n_mean), 
    ub_sample = mean_sample + 1.96 * (sd_sample /((400 * count) ** .5)),
    lb_sample = mean_sample - 1.96 * (sd_sample /((400 * count) ** .5)),) %>% # missing SD here!
  ungroup() 
  
write_csv(resnet_summarised, here("data/model_sim_data/resnet50_summarised.csv"))

resnet_summarised %>%   
   ggplot(aes(x = trial_number, y = mean_sample, color = trial_type)) +
 scale_y_log10() +
 stat_summary(fun.data = "mean_cl_boot") 
 theme_few()
```



# Residual dishabituation 

## preprocessing data 

Residual_dishabituation ~ embedding_distance * violation

Residual dishabituation is the predicted dishabituation magnitude derived from model (1).
Embedding distance is calculated from the eucledian distance between the stimulus used in the background trial and the stimulus used in the deviant trial of the same block.
Violation is the type of violation, which has four levels: number violation, category violation, orientation violation, and identity violation.


```{r}
embedding <- read_csv(here("data/model_sim_data/embedding_distance.csv")) %>% 
  rename(stim_name = `...1`)

replace_name <- function(x) gsub("media/padded_stims/", "", x)
bd_with_embedding <- bd %>% 
  mutate_at(c("background_stimulus", "deviant_stimulus"), replace_name) %>% 
  mutate(
    background_stimulus = case_when(
      grepl("pair", background_type) ~ gsub("(.*)(\\.)", paste0("\\1_pair\\2"), background_stimulus), 
      TRUE ~ background_stimulus
    ), 
    deviant_stimulus = case_when(
      grepl("pair", deviant_type) ~ gsub("(.*)(\\.)", paste0("\\1_pair\\2"), deviant_stimulus), 
      TRUE ~ deviant_stimulus
    )
  )

get_distance <- function(stim_a_name, stim_b_name,  dist){
  if (stim_a_name  == "" | stim_b_name  == ""){
    return (tibble("val" = NA_real_))
  }else{
    return( 
      tibble("val" = dist %>%
      filter(stim_name == {{stim_b_name}}) %>% 
      select({{stim_a_name}}) %>% 
      pull())
      )
  }
  
}

pending_matching <- bd_with_embedding %>% 
  distinct(background_stimulus, deviant_stimulus) %>% 
  filter(!is.na(deviant_stimulus))

matched_val <- pending_matching %>% 
  ungroup() %>% 
  mutate(edist = map2_df(.x = pending_matching$background_stimulus, 
                        .y = pending_matching$deviant_stimulus, 
                        .f = get_distance, embedding)) %>% 
  unnest(edist)


bd_with_embedding <- bd_with_embedding %>% 
  left_join(matched_val, by = c("background_stimulus", "deviant_stimulus"))

```
## residual 

```{r}
bd_with_embedding

bd_with_embedding <- bd_with_embedding %>% 
  mutate(is_first_trial = trial_number == 1) %>% 
  mutate(violation_type_with_background = case_when(
    trial_type == "background" ~ "background", 
    TRUE ~ violation_type
  )) %>% 
  mutate(
    animacy = case_when(
    grepl("inanimate", background_type) ~ "inanimate", 
    TRUE ~ "animate"), 
    
    number = case_when(
    grepl("single", background_type) ~ "single", 
    TRUE ~ "pair"
  ),
  
    pose = case_when(
    grepl("left", background_type) ~ "left", 
    TRUE ~ "right"
  )
    )

bd_with_embedding$violation_type_with_background = factor(bd_with_embedding$violation_type_with_background, 
                                          levels = c(
                                                     "background",
                                                     "pose", 
                                                     "animacy", 
                                                     "number", 
                                                     "identity"))

m <- lmerTest::lmer(
  log(total_rt) ~ trial_number * is_first_trial + (1|prolific_id) + 
    trial_number * is_first_trial * number + 
    trial_number * is_first_trial * pose + 
    trial_number * is_first_trial * animacy + 
    trial_number * is_first_trial * violation_type_with_background + log(block_number),
  data = bd_with_embedding 
)

bd_with_embedding$residual = resid(m)
  


```

## run the model 

```{r}
new_d <- bd_with_embedding %>% 
  filter(trial_type == "deviant")


#Residual_dishabituation ~ embedding_distance * violation

lm(residual ~ val * violation_type, data = new_d) %>% 
  summary()
```


# Archive 


```{r}
eig_hab_dishab <- trial_info %>% 
  distinct(trial_id, fam_duration, violation_type) %>% 
  left_join(eig_sd, by = c("trial_id")) %>% 
  mutate(trial_number = index + 1) %>% 
  select(param_id, trial_number, fam_duration, violation_type, sample_n_mean, sample_n_std) %>% 
  mutate(trial_type = case_when(
    violation_type == "background" ~ "background", 
    violation_type != "background" & (trial_number == fam_duration + 1) ~ "deviant", 
    TRUE ~ "background"
  )) %>% 
  mutate(total_trial_number = fam_duration +1 ) %>% 
  select(param_id, trial_type, trial_number, total_trial_number, sample_n_mean, sample_n_std) %>% 
  group_by(param_id, trial_type, trial_number, total_trial_number) %>% 
  summarise(mean_sample = mean(sample_n_mean)) %>% # missing SD here!
  ungroup() 

eig_hab_dishab %>% write_csv(here("data/model_sim_data/eig_hab_dishab.csv"))
```

```{r}
deviant_only_fit <- trial_info %>% 
  distinct(trial_id, fam_duration, test_type) %>% 
  left_join(eig_sd, by = c("trial_id")) %>% 
  mutate(trial_number = index + 1) %>% 
  select(param_id, trial_number, fam_duration, test_type, sample_n_mean, sample_n_std) %>% 
  mutate(trial_type = case_when(
    test_type == "background" ~ "background", 
    test_type == "identity" & (trial_number == fam_duration + 1) ~ "identity", 
    test_type == "number" & (trial_number == fam_duration + 1) ~ "number", 
    test_type == "pose" & (trial_number == fam_duration + 1) ~ "pose", 
    test_type == "animacy" & (trial_number == fam_duration + 1) ~ "animacy", 
    TRUE ~ "background"
  )) %>% 
  group_by(param_id, trial_type) %>% 
  summarise(mean_sample = mean(sample_n_mean)) %>% # missing SD here!
  ungroup() 

deviant_only_fit %>% write_csv(here("data/model_sim_data/eig_deviant_only.csv"))
```



```{r}
sd %>% 
   filter(trial_type != "background") %>% 
  group_by(param_id, trial_type) %>% 
  summarise(mean_sample  = mean(mean_sample)) %>% 
  ggplot(aes(x = trial_type, y = mean_sample, color = trial_type)) + 
  geom_line(aes(group = param_id), alpha = .3) + 
  #geom_point(position = position_dodge(width  = .2), alpha = .3) + 
  stat_summary(fun.data = "mean_cl_boot", position = position_dodge(width = .2))
```
